modify project for cleaner code extract

+++ database/migrations/2022_10_29_185055_create_transcations_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedBigInteger('order_id');
             $table->enum('gateway',['zarinpall','shaparak']);
             $table->integer('qyt');
-            $table->decimal('amount');
+            $table->unsignedBigInteger('amount');
             $table->string('address');
             $table->string('phone');
             $table->text('description')->nullable();
